fix: Forum folder created and account folder moved into Forum
