+++ main/src/app/(forum)/(account)/a/settings/notifications/page.tsx
@@ -1,5 +1,5 @@
-import SettingsNav from "@/components/Account/Settings_Section/settingsnav";
-import AccountSideNav from "@/components/Account/sidenav";
+import SettingsNav from "@/components/Forum/Account/Settings_Section/settingsnav";
+import AccountSideNav from "@/components/Forum/Account/sidenav";
 
 export default function StgsNotifications() {
 	return (
